requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
typing-extensions
sqlalchemy
pydantic[email]